dmod: Make more robust when used outside default folders
The snippet expansion is now split into two distinct parts: the namespace and the module name.

The namespace expansion looks for a `lib|test|spec` folder and uses any folder below that as the namespace, joined by dots. If no `lib|test|spec` folder can be found, no namespace will be generated.

The module name is generated through the `TM_FILENAME_BASE`, and is basically the same as it was before. But this allows us to generate a `defmodule Config` module in `config/config.exs`.